+++ LLama.Examples/Examples/GrammarJsonResponse.cs
@@ -1,7 +1,7 @@
 ﻿using LLama.Common;
 using LLama.Grammars;
 
-namespace LLama.Examples.NewVersion
+namespace LLama.Examples.Examples
 {
     public class GrammarJsonResponse
     {
